+++ api/polls/models.py
@@ -1,30 +1,19 @@
-from tortoise import models, fields
+from datetime import datetime
 
-from slugify import slugify
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Text
+
+from api.users.models import User
+from core.database import Base
 
 
-class Poll(models.Model):
-    id = fields.IntField(pk=True)
-    title = fields.CharField(max_length=255, unique=True)
-    slug = fields.CharField(max_length=255, unique=True)
-    creator = fields.ForeignKeyField('models.User', related_name='polls', on_delete=fields.CASCADE)
-    description = fields.TextField(null=True, default="")
-    media_type = fields.CharField(max_length=255, default="IMAGE")
-    places_number = fields.IntField(default=0)
-    image = fields.CharField(max_length=500, null=True)
-    created_at = fields.DatetimeField(auto_now_add=True)
-
-    def save(self, *args, **kwargs):
-        self.slug = slugify(self.title)
-        return super().save(*args, **kwargs)
-
-    def update_from_dict(self, data: dict):
-        if data.get('title'):
-            self.slug = slugify(data['title'])
-        return super().update_from_dict(data)
-
-    class Meta:
-        table = 'polls'
-
-    class PydanticMeta:
-        exclude = ('slug', 'creator', 'created_at', 'places_number')
+class Poll(Base):
+    __tablename__ = 'polls'
+    id = Column(Integer, primary_key=True)
+    title = Column(String, unique=True)
+    slug = Column(String, unique=True)
+    creator = Column(Integer, ForeignKey(User.id))
+    description = Column(Text, nullable=True, default="")
+    media_type = Column(String, default="IMAGE")
+    places_number = Column(Integer, default=0)
+    image = Column(String, nullable=True)
+    created_at = Column(DateTime, default=datetime.utcnow)