+++ core/base.py
@@ -1,0 +1,3 @@
+from core.database import Base  # noqa
+from api.users.models import User  # noqa
+from api.polls.models import Poll  # noqa