+++ tests/test_user_registration.py
@@ -1,13 +1,9 @@
-import pytest
-
 from api.users.models import User
-from api.auth.validators import validate_unique_email, validate_unique_username
 
 from core.tasks import send_email
-from core.exceptions import CustomValidationError
 
 
-def test_user_registration(mocker, client, event_loop):
+def test_user_registration(mocker, client, db):
     mocker.patch('core.tasks.send_email.delay')
     data = {
         'username': 'test_user',
@@ -17,7 +13,7 @@ def test_user_registration(mocker, client, event_loop):
     resp = client.post('/api/auth/register', json=data)
     assert resp.status_code == 201
     assert 'id' in resp.json()
-    user = event_loop.run_until_complete(User.get(id=resp.json()['id']))
+    user = db.query(User).get(resp.json()['id'])
     assert user.email == resp.json()['email']
     assert not user.is_active
     send_email.delay.assert_called_once()
@@ -33,7 +29,7 @@ def test_user_registration_invalid_data(client):
     assert resp.status_code == 400
 
 
-def test_user_registration_with_already_existed_email(mocker, client, event_loop):
+def test_user_registration_with_already_existed_email(mocker, client, db):
     mocker.patch('core.tasks.send_email.delay')
     data = {
         'username': 'test_user',
@@ -44,7 +40,7 @@ def test_user_registration_with_already_existed_email(mocker, client, event_loop
     data['username'] = 'another_user'
     resp = client.post('/api/auth/register', json=data)
     assert resp.status_code == 400
-    assert event_loop.run_until_complete(User.filter(email=data['email']).exists())
+    assert db.query(db.query(User).filter(User.email == data['email']).exists())
     assert 'email' in resp.json()
 
 
@@ -61,14 +57,3 @@ def test_user_registration_with_already_existed_username(mocker, client):
     assert resp.status_code == 400
     assert 'username' in resp.json()
 
-
-@pytest.mark.asyncio
-async def test_validation_raise_error_that_email_exists(user):
-    with pytest.raises(CustomValidationError):
-        await validate_unique_email(user.email)
-
-
-@pytest.mark.asyncio
-async def test_validation_raise_error_that_username_exists(user):
-    with pytest.raises(CustomValidationError):
-        await validate_unique_username(user.username)